requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.24.0,<0.25
Pillow>=9.2.0,<9.3
uwsgi>=2.0.20<2.1
